(PC-25661) test(offer): extract offer body tests from offer content
